Makefile: workaround to allow releasing
since 34d519c there is the git-commithash in the firmware-filename,
which broke the current automatic deploymentof releases from buildbot.
This workaround is a hack to restore the old filename-scheme to allow
releasing of kathleen-0.2.0. In parallel we should look for a clean way
for master-branch

setting BUILDTYPE on config.mk decides on the naming-scheme in the
Makefile. Valid values are "unstable" or "release".